requirements.txt
requests
beautifulsoup4
selenium
webdriver_manager
pandas
tabula-py